requirements.txt
hcloud
python-dotenv
pyTelegramBotAPI
rich
pandas
matplotlib